+++ frontend/scripts/react-components/tool/tool-modal/tool-selection-modal/tool-panel/tool-panel.selectors.js
@@ -4,7 +4,6 @@ import { getPanelId } from 'utils/toolPanel';
 import { getDirtyBlocks } from 'react-components/nodes-panel/nodes-panel.selectors';
 import pluralize from 'utils/pluralize';
 import { TOOL_STEPS } from 'constants';
-import { getSelectedContext } from 'reducers/app.selectors';
 
 const getCountriesData = state => state.nodesPanel.countries.data;
 const getSourcesData = state => state.nodesPanel.sources.data;
@@ -208,36 +207,3 @@ export const getIsDisabled = createSelector(
     return !currentSentencePart.optional && !currentSentencePart.value.length > 0;
   }
 );
-
-const getURLParamsIfContext = (params, context) => {
-  if (!context) {
-    return null;
-  }
-  return params;
-};
-
-const getURLSources = createSelector(
-  [getSources, getSelectedContext],
-  getURLParamsIfContext
-);
-const getURLExporters = createSelector(
-  [getExporters, getSelectedContext],
-  getURLParamsIfContext
-);
-const getURLImporters = createSelector(
-  [getImporters, getSelectedContext],
-  getURLParamsIfContext
-);
-const getURLDestinations = createSelector(
-  [getDestinations, getSelectedContext],
-  getURLParamsIfContext
-);
-
-export const getDashboardElementUrlProps = createStructuredSelector({
-  sources: getURLSources,
-  exporters: getURLExporters,
-  importers: getURLImporters,
-  countries: getSelectedCountryId,
-  destinations: getURLDestinations,
-  commodities: getSelectedCommodityId
-});
