+++ frontend/scripts/react-components/tool/tool.js
@@ -5,20 +5,24 @@ import { resizeSankeyTool } from 'react-components/tool/tool.thunks';
 import { getToolLinksUrlProps } from 'react-components/tool-links/tool-links.selectors';
 import { getAppUrlProps } from 'reducers/app.selectors';
 import { getToolLayersUrlProps } from 'react-components/tool-layers/tool-layers.selectors';
+import { getPanelUrlProps } from 'react-components/tool/tool.selectors';
 import toolLayerSerializer from 'react-components/tool-layers/tool-layers.serializers';
 import toolLinksSerializer from 'react-components/tool-links/tool-links.serializers';
+import nodesPanelSerializer from 'react-components/nodes-panel/nodes-panel.serializers';
 
 const { urlPropHandlers: toolLayersUrlPropHandlers } = toolLayerSerializer;
 const { urlPropHandlers: toolLinksUrlPropHandlers } = toolLinksSerializer;
+const { urlPropHandlers: nodesPanelUrlPropHandlers } = nodesPanelSerializer;
 
 const getUrlProps = createSelector(
-  [getToolLinksUrlProps, getAppUrlProps, getToolLayersUrlProps],
-  (toolLinks, app, toolLayers) => ({ ...toolLinks, ...app, ...toolLayers })
+  [getToolLinksUrlProps, getAppUrlProps, getToolLayersUrlProps, getPanelUrlProps],
+  (toolLinks, app, toolLayers, panel) => ({ ...toolLinks, ...app, ...toolLayers, ...panel })
 );
 
 const urlPropHandlers = {
   ...toolLayersUrlPropHandlers,
-  ...toolLinksUrlPropHandlers
+  ...toolLinksUrlPropHandlers,
+  ...nodesPanelUrlPropHandlers
 };
 
 const mapStateToProps = state => ({
