+++ frontend/scripts/utils/translate-link.js
@@ -37,18 +37,7 @@ function translateLink(data, meta, to = 'sankey') {
   const { queryParams, countryId, commodityId, title } = data;
   const params = parseURL(queryParams);
 
-  if (to === 'sankey') {
-    return {
-      type: 'tool',
-      payload: {
-        serializerParams: params
-      }
-    };
-  }
-
   let serializerParams = {
-    selectedCountryId: countryId,
-    selectedCommodityId: commodityId,
     selectedYears: params.selectedYears,
     selectedResizeBy: params.selectedResizeBy,
     selectedRecolorBy: params.selectedRecolorBy
@@ -58,12 +47,34 @@ function translateLink(data, meta, to = 'sankey') {
     const roleQueryParams = getRoleQueryParams(params.selectedNodesIds, meta);
     serializerParams = pickBy({ ...serializerParams, ...roleQueryParams });
   }
+  // ----------------- We need to add the nodes to the nodesPanel state
+  if (to === 'sankey') {
+    return {
+      type: 'tool',
+      query: {
+        ...serializerParams,
+        countries: countryId,
+        commodities: commodityId
+      },
+      payload: {
+        serializerParams: {
+          ...serializerParams,
+          countries: countryId,
+          commodities: commodityId
+        }
+      }
+    };
+  }
 
   return {
     type: 'dashboardElement',
     payload: {
       dashboardId: kebabCase(title),
-      serializerParams
+      serializerParams: {
+        ...serializerParams,
+        selectedCountryId: countryId,
+        selectedCommodityId: commodityId
+      }
     }
   };
 }
